[Patchset 14] trace_processor: implement string pooling class
This CL introduces string pooling functionality to the trace processor
backed by PagedMemory.
This class uses fixed pages of memory with the following
properties:
1. On 64-bit machines, we mmap a 4GB chunk of memory and then use the
offset into this chunk as the StringId.
2. On 32-bit machines, we mmap 32MB chunks (this is because the UI will
malloc anything we mmap so we want to be conservative). The raw pointer
is then used directly as the StringId.

The use of this class inside the trace processor will be carried out
in a follow up CL.

+++ src/trace_processor/string_pool.h
@@ -108,7 +108,8 @@ class StringPool {
     uint8_t* TryInsert(base::StringView str);
 
     uint32_t OffsetOf(uint8_t* ptr) const {
-      PERFETTO_DCHECK(Get(0) < ptr && ptr < Get(kBlockSize));
+      PERFETTO_DCHECK(Get(0) < ptr &&
+                      ptr < Get(static_cast<size_t>(kBlockSize)));
       return static_cast<uint32_t>(ptr - Get(0));
     }
 
@@ -116,8 +117,9 @@ class StringPool {
 
    private:
     static constexpr size_t kBlockSize =
-        sizeof(void*) == 8 ? 4ull * 1024ull * 1024ull * 1024ull /* 4GB */
-                           : 32ull * 1024ull * 1024ull /* 32MB */;
+        sizeof(void*) == 8
+            ? static_cast<size_t>(4ull * 1024ull * 1024ull * 1024ull) /* 4GB */
+            : 32ull * 1024ull * 1024ull /* 32MB */;
 
     base::PagedMemory mem_;
     uint32_t pos_ = 0;
